Updated release scripts.
Thanks to xet7 !

+++ releases/release-sandstorm.sh
@@ -1,36 +1,34 @@
-# Usage: ./release.sh 3.95
+#!/bin/bash
 
-# Delete old stuff
-#cd ~/repos/wekan
-#./releases/release-cleanup.sh
+# Release Sandstorm Wekan.
 
-# Build Source
-#cd ~/repos/wekan
-#./releases/rebuild-release.sh
+# 1) Check that there is only one parameter
+#    of Sandstorm Wekan version number:
 
-REPODIR=/home/wekan/repos
-WEKANDIR=/home/wekan/repos/wekan
+if [ $# -ne 1 ]
+  then
+    echo "Syntax with new Sandstorm Wekan version number:"
+    echo "  ./release-sandstorm.sh 5.10.0"
+    exit 1
+fi
 
 # Ensure sudo access
 sudo echo .
+
+# Delete old temporary build directory
 rm -rf ~/repos/wekan/.meteor-spk
+
+# Start and update local Sandstorm dev version
 sudo systemctl enable sandstorm
 sudo sandstorm start
 sudo sandstorm update
-# Build Sandstorm
-#cd $REPODIR
-#rm -rf $WEKANDIR
-#git clone git@github.com:wekan/wekan.git
-#cd $WEKANDIR
-#sudo n 12.21.0
-#sudo mkdir -p /usr/local/lib/node_modules/fibers/.node-gyp
-# Build Wekan
-#./releases/rebuild-release.sh
-#cd .build/bundle/programs/server
-#npm install node-gyp node-pre-gyp fibers
-cd $WEKANDIR
-# Build Sandstorm
+
+# Build Sandstorm Wekan
+cd ~/repos/wekan
 meteor-spk pack wekan-$1.spk
+
+# Publish Sandstorm Wekan to exprimental App Market
 spk publish wekan-$1.spk
+
+# Delete old temporary build directory
 rm -rf ~/repos/wekan/.meteor-spk
-#scp wekan-$1.spk x2:/var/snap/wekan/common/releases.wekan.team/
